add functionality to single script
